feat(FLCE): expose `accum_dtype` for hf model monkey patch (#851)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
This PR is a follow-up to #830, exposing `accum_dtype` option for monkey
patch functions.
All bf16 convergence tests related to fused linear cross entropy are
also enforced to run with `accum_dtype=torch.float32` for numerical
stability.

Related: #512, #742, #827, #850 
<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [ ] run `make test` to ensure correctness
- [ ] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence

+++ test/convergence/bf16/test_mini_models.py
@@ -926,7 +926,7 @@ def run_mini_model(
     for i in range(num_steps):
         batch = next(loader_iter).to(model.device)
         optimizer.zero_grad()
-        output = model(**batch)
+        output = model(**batch, accum_dtype=torch.float32)
         output.loss.backward()
         optimizer.step()
         print(f"Step {i}, Loss: {output.loss.item()}")
